Script

+++ autolister.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-# Check if the input file is provided
-if [ -z "$1" ]; then
+# Check if the input and output files are provided
+if [ -z "$1" ] || [ -z "$2" ]; then
     echo "Usage: $0 domains.txt output.txt"
     exit 1
 fi
@@ -20,8 +20,11 @@ while IFS= read -r domain; do
     echo " "
     echo "Running Sublist3r for ===> $domain"
     sublist3r -d "$domain" -v -t 150 -o temp_subdomains.txt
-    cat temp_subdomains.txt >> "$output_file"
+    
+    if [ -f temp_subdomains.txt ]; then
+        cat temp_subdomains.txt >> "$output_file"
+        rm temp_subdomains.txt
+    else
+        echo "No subdomains found for $domain"
+    fi
 done < "$input_file"
-
-# Remove the temporary file
-rm temp_subdomains.txt
